CLEANUP: refactored max_element_bytes code.

+++ engines/default/default_engine.h
@@ -61,7 +61,7 @@ struct engine_config {
    uint32_t max_set_size;
    uint32_t max_map_size;
    uint32_t max_btree_size;
-   size_t max_element_bytes;
+   uint32_t max_element_bytes;
    bool   ignore_vbucket;
    char   prefix_delimiter;
    bool   vb0;
